Autoupdate of 2 submodules
*hardware/interfaces updated to a886a5c4bea647717d65d78f7b5db1a8b83daf4e
Merge changes from topic "fix-aidl-warnings"

* changes:
  aidl: Set explicit default values
  aidl: Set explicit default values
  aidl: Set explicit default values for enum-type fields
  aidl: Set explicit defaults to enum-type fields

*system/tools/aidl updated to 308bbe08b3c07a2113e7130ec75ebe5cd8fb93a0
checkapi: Allow default values for enum-type fields

We recommend to set explicit default values for enum-type fields. But
--checkapi complains about setting a default value even if the value is
zero which is compatible with ""(empty) for enum types.

For aidl_interfaces with versioned snapshots, we can't fix
-Wenum-explicit-default because --checkapi sees it as an incompatible
change to set the default value in the new while the old doesn't set
the default.

Bug: 179853398
Test: aidl_unittests
Change-Id: I11b773f7459100b23944164192379369e305783c