Add the script data files
The data/ directory includes the `static` and `templates`
directories which is used as the source for the generated Bazel
environment. Files placed in the template directory are formatted and
staged, and eventually synced to the source tree. Thus, they are stored
in the same directory hierarchy as the source tree. Files placed in the
static directory are copied into the source tree.

Test: N/A
Change-Id: I5adb3cd95e3142030287132b6ebc88d21da4f15c