Autoupdate of 4 submodules
*tools/data-binding updated to 1759f730db78a70a240f2471f7a981ac5209e422
Add publishing block for publications in databinding

Bug: 150132264
Test: n/a
Change-Id: I61396fbaecd89c3f61cb8719d7b9ca59e968a178

*tools/adt/idea updated to 7c76ba65b60c25c402601d09fb563525d01feddf
Add publishing block for publications from IDE tests

Bug: 150132264
Test: n/a
Change-Id: Id8f3ecb158f7d92ac3304c79840e055e6c4e0a13

*tools/analytics-library updated to f19b739319d62db79b3b649725e2d118fd5fee1c
Add proto emum value for disable automatic component creation flag

Bug: 150132264
Test: n/a
Change-Id: I47fbade195ff7701a3bb010894204d32d51fec10

*tools/base updated to 59abfdbcd12a4a27ae8d3c6bfbaaed818a1faf28
Single variant publishing for library and application projects

With this change, we add the new publishing dsl for users to publish a
single variant to a component. For library projects, the variant is
publish as an AAR artifact. For application projects, the variant is
published as an APK or AAB. Note publishing the variant as APK and AAB
to the same component is not supported. By default, we publish the
application variant as AAB.

With the new publishing mechanism, we are not going to create software
component automatically and the component name for APK/AAB is same as
the variant name instead of something like variantName_aab or
variantName_apk.

In order to offer a better upgrade experience for users, we still create
"existing" components if users don't opt in this feature(they don't use
the new dsl and don't use the new introduced flag to disable automatic
component creation). When developers don't opt in this feature, we raise
warnings about the behavior change so that they can migrate to use new
publishing dsl before the support for old behavior is dropped
completely.

Bug: 150132264
Test: existing + LibraryPublishingTest + AppPublishingTest
Change-Id: Iabea55a7692c8736f13c6337c61e6a79401489da