Autoupdate of 2 submodules
*prebuilts/r8 updated to ed57ab1e86361ce0e8c12b88b481378eddcd1143
Update D8 R8 to 3.1.0-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.1.0-dev
Test: ./gradlew check
Bug:

Change-Id: I36ef9625a49e0199da5df191eca6ced1888d70ee

*tools/base updated to 70d89b5918fdf7facdacc9ccad5dceeb0bf0da7d
Update test expectations after L8 use R8 full-mode shrinking

Test: CoreLibraryDesugarTest
Bug: N/A

Change-Id: I4e003f6722df2e8787db80d38800a616a404623f