Change libvixl symbol visibility.
Make all symbols protected for shared library and hidden for
static library. This reduces the size of the shared library
by ~44KiB on arm and ~78KiB on arm64, configured with both
arm and arm64 target. Binaries that link libvixl statically
are also a lot smaller, depending on the used libvixl parts.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 186902856
Change-Id: I5732117c4f2a564a032fa2642957efb63390035b