Autoupdate of 2 submodules
*tools/adt/idea updated to bc5dd8b5e62ac9952ccfe13f644f60e9e0fbdd2a
[AUA] Support for removing implementation properties

A number of properties of BaseExtension and some of its contained (in
the Dsl sense) classes are no longer available by exported interfaces.
This doesn't immediately matter for Groovy, though even the
implementation classes are expected to be separate by AGP 8.0.0, but it
has immediate ramifications for KotlinScript, where the build scripts
will not compile if they have references to implementation-class
properties that are not present (or write references to read-only
properties) in the interfaces.

Implement an upgrade RefactoringProcessor, with the slightly different
feature that its necessity is not characterized by a region, in that we
cannot from the Upgrade Assistant easily tell whether properties are
expressed in Groovy or KotlinScript build files, so we need to continue
to make this a mandatory upgrade if the region [current,new) intersects
a *region* of version space, rather than just a point.

This version restricts itself to removing only those properties
introduced by the 4.x templates for dynamic-feature modules, as a useful
staging point with a test case before the fuller implementation.

Bug: 186647828
Test: added
Change-Id: I5a07a3b63f15fb17ff73ce6d3283d49f2c9a9d89

*tools/analytics-library updated to d9a4d5f69a0c1b46a23172a706ecf592f633c2cd
Update studio_stats.proto

Logs-Approval: cl/372345029
Bug: 186647828
Test: existing
Change-Id: Ic2d5a13a30b04deff75b61f30bfad74add1a0470