Autoupdate of 2 submodules
*tools/adt/idea updated to 78954662f49dffa371b58d85bd1455e42529284f
Stop sending information that we already have

There's no need to send the device api level and package name in layout
event messages, since we already have that in studio.

Test: updated existing tests.
Bug: none
Change-Id: I3477c77db41407e632062185da5f42a778bd63a6

*tools/base updated to 24e788a286211ce1ba58b1078fddd23ed21a3442
Stop sending information that we already have

There's no need to send the device api level and package name in layout
event messages, since we already have that in studio.

Test: updated existing tests.
Bug: none
Change-Id: If86fdd37aafff8d281ea4bc49929fe6b7ed414d1