Do not rely on gradle deps on ModuleClassLoader
When jetifier is enabled, a project could be using classes in the
androidx namespace but not have the androidx dependencies. This break
the heuristic to determine which adapters should be used in
ModuleClassLoader.
This CL removes that check and creates the adapters based on the
namespace of the RecyclerView being loaded.

Fixes: 173155464
Test: Manually verified in a project with support deps and jetifier
enabled.

Change-Id: Ib7e99f296613bae875d16ad746003d7c6e2c29b7