Fix Debian CI build.
Debian runs reprotest, which changes around some things about the
environment. Among other things it changes the uname, which confuses gn
about the architecture we are running on. By using DEB_BUILD_GNU_CPU we
are stable against that.

Also add the yaml config to run the CI on salsa.debian.org, so we do not
need to add a file to the repo there.

Test: builds on salsa.debian.org: https://i.imgur.com/1yNVqTy.png

Bug: 184929946
Change-Id: I8304ee4878f4853a397fe98d21e50f6998ef14bf