Re-enable ignored test
Previous, this test would flake occasionally due to an error:

```
kotlin.NotImplementedError: An operation is not implemented.
  at StubTestAppInspectorMessenger.getScope()
```

It looks like the cause of that exception was fixed on April 16, as
that method was implemented for a different change, so we can safely
re-enable this test now.

I ran this test with runs_per_test=1000 locally to test for flakes.

Fixes: 184467010
Test: This CL removes an @Ignore
Change-Id: I03921191c51a798a5d52179462d74dea3404e821