Autoupdate of 3 submodules
*tools/data-binding updated to 4896c5af80e6ed485cf755e7d5ce556502a544d1
Extract GradleBuildInvoker interface

This is to integrate it with ApkProvider and remove deprecated `outputs`
properties from IDE models.

Bug: 37797956
Test: n/a
Change-Id: If428139e71aa83c05deaee01a1464858f8c4945a

*tools/adt/idea updated to 63054ca4d5f8f587f1825780cb43426f59094583
Extract GradleBuildInvoker interface

This is to integrate it with ApkProvider and remove deprecated `outputs`
properties from IDE models.

(1) Replace GradleBuildInvoker.Request with Kotlin data class and
    a builder

(2) Move `executeTasks`s `listener` parameter back to `Request`

(3) Rework some of `GradleBuildInvokerTest` so that they test at least
    something.

Bug: 37797956
Test: n/a
Change-Id: I2f7541660e8b4b62909a3f882ed4c4c1084a82b9

*tools/vendor/google updated to 54c038762965ce0a18e239a7770c2eb84866a036
Extract GradleBuildInvoker interface

This is to integrate it with ApkProvider and remove deprecated `outputs`
properties from IDE models.

Bug: 37797956
Test: n/a
Change-Id: Id9381b7d0e34891eb73a2e2ceb474adbe50ca34c