Autoupdate of 2 submodules
*external/grpc-grpc updated to fcecb51d25341dc00a016f42da3e442741d259ae
[GRPC] Dup the provided fd for full ownership

Bug: 186054748
Test: manual; used logs to verify the fix
Change-Id: Ic3e387f8c5b222e0cfb35718921266f7ef7cdc5a

*tools/base updated to f285e3e2234e86e00b8376456603f4a34ebfb903
Enforce the ownership of gRPC socket's fd

As the gRPC library would make a copy of the fd,
the agent should tak the full control of the fd's
lifetime and close() it when needed.

Logs the close() call if there's an error.

Also makes the DaemonStatusChanged callback configurable
because sometimes it should be thrown away after execution
but sometimes it shouldn't.

Bug: 186054748
Test: manual; used logs to verify the fix
Change-Id: I92611bb09cbbe01ef8500da2618ffd2bca2f47d0