Autoupdate of 2 submodules
*packages/modules/Virtualization updated to 24abe6ae4b3c0a1133bc0e3cd1fcff9ecee19184
Merge changes from topic "microdroid_launcher"

* changes:
  Put ro.product.cpu.abi
  Add microdroid_launcher

*system/sepolicy updated to 6645ad3b1f0edf78a2e0cd8ecf17665d7c108d9d
Add rules for microdroid_launcher

Microdroid_launcher is an executable in microdroid. It's role is to load
a shared library in an APK that is shared from the host Android and
execute it by calling an entry point (android_native_main) in it.

For now, it is executed from shell, but will eventually be executed from
a binder service (which also is running in microdroid) called
microdroid_manager.

Bug: 188513012
Test: atest MicrodroidHostTestCases
Change-Id: I150a958c1ed0e3e960f4b4b577e808e54e898644