Push intellij.cidr.debugger.plugin dependency to android-ndk
Transitive dependencies between platform plugins are not supported
(in our case it would be more work than in gradle-intellij-plugin
https://github.com/JetBrains/gradle-intellij-plugin/commit/ba9200190a0257325b4b47d3cd8106a48e89b873).

But instead of adding extra dependencies to test modules, it seems
reasonable to add a dependency in the android-ndk module, i.e. just
the first consumer of the cidr plugin has to explicitly list the
transitive closure.

Change-Id: Ibd20baf7cb8a9bf3cdf629572b7e876b9ddebb1f