[AUA] Fix a test
One of the two test cases failing in this bug report is clear: if the
latest AGP version known to the system is also one of the known versions
given, it will not be presented twice.  Adjust that test to deduplicate
appropriately.

The other test failure is more mysterious.  Write the test to be more
explicit about what it is expecting, so that in the event of the
expected failure on Mac it should be easier to see what is going on.

Bug: 189890821
Test: these are tests
Change-Id: I8402b72b7fe53bc2556fef0773b0e29c09b8f4ee