Move the ndk_dmg config into its own file
The repository/repository_common/ndks.yaml configuration is imported by
multiple files, causing the ndk_dmg config to be duplicated.

Duplication of the ndk_dmg config causes the same DMG to be notarized
and staged, which ultimately fails in lorry when the contents of the
DMGs do not match.

Moving the config to its own file prevents such duplication.

Bug: 190546685
Test: N/A
Change-Id: Ia1c8994c12a9683892f3736348d4e7e6eef280bc