odrefresh: enable loading of system_server image files
Fix check in ValidateBootImageChecksum() to account for image files
with multiple components which exists for boot-framework when
compiling on device.

Update logic for only_load_system_executable to be
only_load_trusted_executable and treat /system and the ART APEX data
directory as trusted.

Add test to check .art,.odex,.vdex files derived from the
system_server classpath are mapped when the ART module updates.

Add test to check .art,.oat,.vdex files for the boot class path
extensions are present in the zygote processes.

Bug: 180949581
Test: atest com.android.tests.odsign.OnDeviceSigningHostTest#verifyGeneratedArtifactsLoaded
Change-Id: I3114fc6393402d8da2eb16ba756ab5fab713dc20