Autoupdate of 2 submodules
*tools/base updated to d137ed6c2bd4ae340f13eee9a6463608919756ee
Allow precompiled header to work

The issue was that CMake was emitting a line in
compile_commands.json for the generation of the .pch file.
Android Studio just expects source to object lines so it
caused a red bubble error. This change ignores lines from
compile_commands.json.bin that generate a .pch file.

Bug: 187448826
Test: Added integration test for build part
Change-Id: I11b4b250784513b96ccb69f5c2dea1ed5c6def73

*tools/vendor/google updated to acadd84beaa0536bb4999cdbb2769be7337c7ca0
Allow precompiled header to work

The issue was that CMake was emitting a line in
compile_commands.json for the generation of the .pch file.
Android Studio just expects source to object lines so it
caused a red bubble error. This change ignores lines from
compile_commands.json.bin that generate a .pch file.

Bug: 187448826
Test: Added integration test for build part
Change-Id: I0d7c6cfdac69f251b35b5e2cea76222ad2af7a1c