Add more java doc for variant publishing control
As we have to wait the feature goes into beta to update the DAC page
https://developer.android.com/studio/build/maven-publish-plugin, it
would be nice to have more java doc including some code samples to
introduce this new feature to users. And we could have a brief
note in our release notes and point users to the java doc.
Once we are able to update the DAC page, we could simplify the java doc.

Bug: 184817011
Test: n/a
Change-Id: I5a44c00e8a6cb1742a6f8390f473e22c3e2dce2c