symtab-reader: add support for binaries compiled with CFI
Control-Flow-Integrity (CFI) when enabled in clang built binaries
introduces an indirection when looking up ELF symbols. For DSO, the
symbol table (.dynsym) will still contain the symbols, but additional
symbols with suffix .cfi will be added to the full .symtab.
Unfortunately, the DWARF debug information refers to CFI symbols by
address to the .cfi suffixed variants as they point to the actual
implementation.

When the dwarf reader is determining whether to suppress variable or
function declarations, it does so by identifying if there is an
associated ELF symbol at the given address read from DWARF. Unless we
know about the alternative address, this will fail and the type
information will be suppressed.

Hence add the .cfi symbol values to the lookup map to associate the
address of them with the corresponding publicly exported symbol.

   * src/abg-symtab-reader.cc (symtab::load_): use new
   add_alternative_address_lookups method.
   (add_alternative_address_lookups): New method.
   * src/abg-symtab-reader.h (add_alternative_address_lookups): new
   function declaration.
   * tests/data/test-read-dwarf/test-libaaudio.so: New test data.
   * tests/data/test-read-dwarf/test-libaaudio.so.abi: New test data.
   * tests/test-read-dwarf.cc: New test case.

Bug: 190554910
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Ic2f8455894a1dea6dad36485807ebe8e9abb3d8d