188871862: Unused resources with import aliases
When android.nonTransitiveRClass=true, the R.jar for library
dependencies are missing, which means that in downstream modules, the R
class reference cannot be resolved.

This CL adds a workaround for that: when a compilation unit has aliased
imports, and a symbol cannot be resolved, lint will check the import
statements for resource references there.

Bug: 188871862
Test: Unit test included
Change-Id: I90e56a97a81806b6fb8f9e21189f8608d08282cf