Make multi-configuration type resolution consistent with compiler
If a layout with multiple configurations is not consistent about what
type is associated with an ID, e.g.

```
layout/activity_main.xml
<TextView id="inconsistent_view" />

layout-land/activity_main.xml
<Button id="inconsistent_view" />
```

the compiler compromises by typing that field as the top-level "View"
class.

However, the IDE was always using the main target (so, here, we would
have said the type was "TextView" and not "View"). This was causing
projects to get a mismatch between the IDE's auto-complete type and
what the compiler was spitting out.

Bug: 154531811
Test: Added
Change-Id: I2a708b4f73739f4fa8bf8ac23aa1ecffe2871c30