Autoupdate of 3 submodules
*tools/adt/idea updated to e99d1d8d1a5ab5be55080d79c624125187ccef46
Adds support for cancellation to Deployment

This commit allows users to cancel a deploy through Studio's
UI widget.

Test: manual via setting up a sleep in deployer and hitting Cancel
Bug: N/A
Change-Id: I233e2fef2c09ff049e5e0427a6f9a20be7e55833

*tools/base updated to 35d9db196f220801ee73da756cf434f50c98351a
Adds support for cancellation to ApkInstaller

This commit checks for user cancellation and cancels the ongoing
install when the user attempts to stop deployment.

Test: TaskRunnerTest, DeployerRunnerTest, DeployTaskTest
Fixes: N/A
Change-Id: I065be88012112fb5c42753c7e8136cf01db32046

*tools/vendor/google3/blaze updated to 21b7888d0a7c79014c4c890f0b5ba8b66704e540
API change to add support for cancellation to Deploy

Test: N/A
Bug: N/A
Change-Id: Ia2ac6d0401ff8918164a773060a774e4d579c289