Update D8 R8 to 3.1.10-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.1.10-dev
Test: ./gradlew check
Bug:

Change-Id: I738112a944b2f3b41e7d7c81c2f359349caa29c7