Autoupdate of 2 submodules
*tools/analytics-library updated to 00cff30b4af5b2bcc815618cf1f0751ce8c38047
add setVariantConfig to variant API

Test: Existing
Bug: N/A

Change-Id: I9ccef5944ed5e3309d0b23f12e3d8715b7da33b0

*tools/base updated to bd5fcb3ff4a8caefc6920cfce7bd862883114d30
add variant API to set signing configuration.

The Variant API SigningConfig had properties to set the various
enable flags related to V1/V4 signing but nothing related to the
keystore to use.

Although it is not offered to reset DSL information directly, a
convenience API is provided to identify the name of the DSL
SigningConfig to be used to sign the variant's artifacts.

Test: added recipe
Bug: N/A
Change-Id: I49bb71513d6988026f85cb3827ebb6aa0c33965f