Autoupdate of 3 submodules
*tools/data-binding updated to 2d6b4d20ca5b6acf60b93f998205a316d8d4d621
Add support for tools:viewBindingType

This new attribute allows developer to change the type of the view for
ViewBinding.
This is a critical use case for large screen layouts where different
configurations have different sub types for a common base class and this
configuration parameter lets user handle those cases better by
specifying the parent type.

These values are not validated aside from a simple check across
configurations because view binding does not know about classpaths.

This attribute is not supported for `<include>` tags since they refer to
a generated class when the included layout is a binding. We could
potentially support include cases when the layout is not a binding
layout as a followup CL (if there is demand).

Bug: 154531811
Test: ViewBindingTypeTest, ViewBinderGenerateJavaTest

Change-Id: I22456a1b4e133a24d51f155aae2ad02149e02a74

*tools/adt/idea updated to 78940ff188ab0c742143cd01efff6337f12441ef
Add support for tools:viewBindingType attribute

This CL just adds support for detecting the new viewBindingType
attribute and generating correctly typed fields based on its value.

The fact that invalid cases aren't linted yet is still a big gap, but
that will be handled in a followup CL.

Bug: 154531811
Test: Added
Change-Id: Ie5d44f4c2cc4e58a40fe3352b51de66d85b70723

*tools/base updated to 8da059c23eed5c9358932ef2f3f3e96c76c30c92
Add new viewBindingType attribute to SdkConstants

Bug: 154531811
Test: N/A
Change-Id: I44fe22ceef19f757ccde97704f9b319180a78332