[Dsl] Move KotlinScript support to its own module
At this point, the only things depending on the Kotlin plugin in the Dsl
module are the Kotlin parser / writer files themselves.  Move them to
their own module, and adjust the dependencies of the two resulting
modules: the .dsl module depends on the Groovy plugin; the .dsl.kotlin
module depends on the Kotlin plugin.  This requires making a small
number of utility routines public rather than internal.

In order to maximize the chances of early detection of problems, keep
the gradle-dsl.xml plugin fragment unchanged, including its declaration
of the KotlinDslTransformerFactory implementation of the
GradleDslTransformerFactory interface.  This means that any non-trivial
user of the .dsl module (anything instantiating GradleDslFile) will need
also to have a dependency on the .dsl.kotlin module, or else the
extension point will fail to initialize.  (I think this is preferable to
a proper modularization, which might lead to accidentally depending on
or distributing the Groovy support only.)  Adjust dependencies of other
modules to incorporate the .dsl.kotlin module where necessary.

Fixes: 190485237
Test: existing
Change-Id: I2484d19fecd1fec1ca18e0e6bff95967eebead20