Android Platform annotations support in lint
Lint has a number of checks around annotations -- threading, number
ranges, resource id's, and so on. These are all based on the
androidx.annotation package.

However, in the Android platform source code, these annotations are in
a different package (android.annotation), since androidx can depend on
the platform but not the other way around. As part of the build, the
extracted platform annotations are converted to androidx annotations,
such that when analyzing user code, lint can uniformly use androidx
annotations anywhere and apply constraints from both the Android SDK
and from the user's own code and libraries.

However, lint also runs on the Android platform sources themselves
during development and in gerrit. The annotations-based checks
currently do not enforce anything since they don't recognize the
android.annotation package.

This CL changes that; lint will now also internally convert all
platform annotations into the corresponding androidx annotations before
analysis. And to make sure that this is working properly, it also adds
a new test mode which re-runs any unit test which mentions
androidx.annotation, but (a) replaces all references to
androidx.annotation in the source code with android.annotation, and (b)
replaces the support library jar with platform annotations instead
(which is done as a bytecode rewriting of the androidx ones instead of
checking in a whole new copy.)

Test: ~150 existing unit tests dynamically modified to check for
   the same errors using android.annotation in the source instead
Change-Id: I7cf1bdf639b063dd1c10575bf31e28d96ed20a60