Autoupdate of 2 submodules
*tools/adt/idea updated to e75cca79a21875f262e85d2e992c3d0fc84c5e98
Fix patterns in the IDE snapshots.

Also, remove DefaultNdkVersion from the Snapshot files as this is a default value that is determined by the AGP.

Bug: N/A
Test: updated
Change-Id: Ifa7cbd5ca68da573923ae5745f99161fd8bcb6b8

*tools/vendor/google updated to 960212d0ff3787aaad80dddbb2361a8c0dc44c9a
Fix patterns in the IDE snapshots.

Also, remove DefaultNdkVersion from the Snapshot files as this is a default value that is determined by the AGP.

Bug: N/A
Test: updated
Change-Id: I06f9f9fa004c3d031c4996cb5d6e5efa4d6d2300