tp: overhaul thread_time_in_state to display mcycle counter
This CL changes the time_in_state metric events to output mcycles
directly for the UI to display.

Previously, we use to display mcycles/sec because the UI did not have
a delta mode. Now we have that, we can use the UI support
for this instead of hacking this at the metric level

Change-Id: I16b6f98defcf56db5e8ea70e40062ca02d02f6f6
Bug: 190507133