Autoupdate of 2 submodules
*tools/analytics-library updated to 35ffdb0335702dece4065b0975ab8fe0f5c48246
Add ANDROID_LINT_TEXT_OUTPUT to analytics_enums.proto

Bug: 191897708
Test: existing
Change-Id: I4cd2c63e9f45af35c61758face0be3bc379da2ab

*tools/base updated to 2ea2481f2afc4382eb3dd55e76e0cd125f9b41db
Add AndroidLintTextOutputTask

Previously, if the AndroidLintTask was UP-TO-DATE, AGP would not print
any lint issues when the user ran `./gradlew lint` (which wrongly
suggests that there are no lint issues).  With this change, AGP prints
the appropriate text output even if the AndroidLintTask is UP-TO-DATE.
To achieve this, this change adds a new AndroidLintTextOutputTask task,
which is never UP-TO-DATE. This task (1) prints any necessary text
output, and (2) fails the build if there are lint errors.

This change also prevents AGP from printing lint output if it starts
with "No issues found" or if it contains "0 errors, 0 warnings".

Fixes: 191897708
Fixes: 158259845
Test: LintTextOutputTest, AndroidLintTextOutputTaskTest
Change-Id: I002903f8539345f832518f422fe7072de7de54b4