Autoupdate of 2 submodules
*tools/analytics-library updated to fe7a9b3207e5156bc619d3daddd196f3d2e67c17
Add enums for CheckJetifier task and related StringOption

See Change Ib80338cc for context.

Bug: 167700246
Bug: 117948942
Test: AnalyticsUtilTest
Change-Id: I93ce776e7b1ecf59bd4152466f87a55f8e19a3ff

*tools/base updated to 8d9aeb5cb2d452dfb223a745032fec017136201a
Add CheckJetifier task to see if Jetifier is needed

This is done by checking whether the dependency graphs contain any
legacy support libraries.

To use this task on the command line, run:
    ./gradlew checkJetifier
and the analysis results will be shown on the console.

Note that this task is currently not compatiable with configuration
caching due to https://github.com/gradle/gradle/issues/12871. Therefore,
if the user enables configuration cache in gradle.properties, they will
need to run the task with --no-configuration-cache to override that
setting.

This task can also be called via the Build Analyzer UI. Build Analyzer
can pass the following property to get the results via a file and
present them in a more user-friendly format:
    ./gradlew checkJetifier --no-configuration-cache
        -Pandroid.injected.checkJetifier.resultFile=some-file-path

Bug: 167700246
Bug: 117948942
Test: New CheckJetifierTaskTest
Change-Id: Ib80338ccb17482dd05f8750cab82ff675bb65675