[Dsl] Remove SigningConfig special-cases in ReferenceTo
Now that our interfaces allow us to express this, rewrite ReferenceTo in
terms of GradleDslElementModel: both the context and the reference
itself are to GradleDslElementModel objects, which SigningConfigModels
are.  Rewrite the existing caller of the special-case SigningConfig
ReferenceTo constructor in terms of a properly-context-sensitive
ReferenceTo constructor, and delete the special-case constructor.

Bug: n/a
Test: existing
Change-Id: I0992e14f5af9324a3e5ce1030020d4e1b7ac8b5b