[AUA] Inline and remove AGP_UPGRADE_ASSISTANT flag
I don't think we're going back to the old world.  Remove all the code
that was protected by AGP_UPGRADE_ASSISTANT flag checks, except for some
tests.  Adjust the RecommendedPluginVersion upgrade integration tests;
the test boundary is now whether the Upgrade Assistant was invoked
appropriately given (mocked) user input, relying on our separate Upgrade
Assistant tests themselves for the correctness of operation of the
Upgrade Assistant.

The ForcedGradlePlugin upgrade tests are a little harder to handle with
minimal code change, given that they do not invoke the
straightforwardly-mockable tool window service.  For now, stub the one
test that this affects out, but note that in fact all those tests need
rewriting as they inspect instances of classes which are no longer
relevant to the upgrade process.  A subsequent CL will address this.

Bug: n/a
Test: existing
Change-Id: Id98515280714f16a9e98bb2105fa9e6f6575423e