Autoupdate of 3 submodules
*tools/adt/idea updated to 8b6d703acecb0076523ea3b48da0f9c4042a4a0e
Move maven repo building to build time

Before this change maven_repo would generate a list
of files to be included in a maven repo together with
metadata for them (classifiers) at build time. At
runtime (tests) this file was taken to resolve a maven
repo layout, and then create symlinks for it.

This change moves the resolution phase to build time
and the generated file is already in repository layout.

This now only allows for the new incoming maven rule set
to work but simplifies several things:

. RepoLinker modules are gone, as it is now a simple utility
  function in test utils
. Tests do not have to depend on maven libraries to know how
  to layout a repository.
. The manifest files are simple _zipper argument lists
. The maven_repo_zip rule became an optional output of the rule

Test: existing tests + updated RepoLikerTest
Bug: 190268212
Change-Id: I3fd93020e81f83e881d67a412b82ab1570dce3c9

*tools/base updated to ba0fb56f21db11368f96f7d8000b4cfbcea48acc
Make new repo rule be compatible with the old one

maven_repository now outputs the same files and
providers as maven_repo.

This change also uses the new rule to build the
hybrid gradle AGP build.

Test: this is exercised via the new hybrid build
Bug: 190268212
Change-Id: I6fa7aea97a309ca3ee4c6e6a0239df77e80b504b

*tools/vendor/google updated to f48c1c3afee9451474e9720d5d87461913dffec6
Move maven repo building to build time

Before this change maven_repo would generate a list
of files to be included in a maven repo together with
metadata for them (classifiers) at build time. At
runtime (tests) this file was taken to resolve a maven
repo layout, and then create symlinks for it.

This change moves the resolution phase to build time
and the generated file is already in repository layout.

This now only allows for the new incoming maven rule set
to work but simplifies several things:

. RepoLinker modules are gone, as it is now a simple utility
  function in test utils
. Tests do not have to depend on maven libraries to know how
  to layout a repository.
. The manifest files are simple _zipper argument lists
. The maven_repo_zip rule became an optional output of the rule

Test: existing tests + updated RepoLikerTest
Bug: 190268212
Change-Id: Ie899fb6186da4cbd118d8c48f91029455440035c