Fix getting APK location for app indexing testing
With AGP 4.x+ `IdeAndroidArtifact.outputs` property is not populated
as it is not supported by the newer AGP versions. With this change we
retrieve the location via a helper method in `GradleProjectSystem`.

Unlike the old implementation we do not return the last listed APK in
the case when multiple APKs are produced and instead fail the operation.

Bug: 194058522
Bug: 147586218
Bug: 37797956
Test: n/a (manually tested that APK is located but network error on upload)
Change-Id: I3788d1c20b6b220f80b67ea9851d537410a704f8