Adds stub transform for LiveEdit instrumentation
Creates a stub transform which inserts a prefix on all methods of a
target class. The prefix adds a branch which invokes a Java bytecode
interpreter to execute a new method body specified in a payload.

Entry into the branch is controlled by a boolean flag in the
InterpreterStubs class, which is set when a bytecode payload is present
for a given method (keyed on class name/method name/signature).

Bug: None
Test: Manual
Change-Id: I245319390c7a721de91322f45bfc50d72546f4a8