[AUA] Set the buildModelContext's AGP version
Setting the buildModelContext's AGP version has no effect except when
writing Dsl for newly-created properties, and even then only if there
are applicable version constraints in the Dsl parser tables.  For the
most part this is not exposed in any of our current tests, but as the
tables get marked up with version constraints corresponding to operator
removal, this will be necessary for the Upgrade Assistant to emit
correct build Dsl.

Bug: 194192720
Test: existing
Change-Id: I815a73f9494d92084bbc6263f18c1bec1494c3c7