ART: Implements a test option of forced generation of .cfg
This patch introduces the option to dump .cfg for any type of test. If
the "--dump-cfg <path/graph.cfg> option" is present, then it generates
a .cfg following the same logic used for the "checker" tests but
forcing it for any type of test. This option works for both host and
target.

Before the artifacts are wiped, the generated .cfg file is copied to
the specified host path.

Note that it will only dump a .cfg if the compiler type is set to
"optimizing". Otherwise, it will throw a warning.

Test: ./art/test/run-test  --host --simulate-arm64  --optimizing \
--dump-cfg <path>/graph.cfg  <test>

Change-Id: I034b13b401bd1dee8d871297da8d09e807a35fd1