Autoupdate of 2 submodules
*build/blueprint updated to 030150d8f9d164783ea661f07793c45198739cca
Add function to retrieve module by name.

This function should only be used in corner-cases where it is not
possible to retrieve the module by traversing a dependency graph; there
are no guarantees about which variant is returned, and the function will
eagerly panic if something is amiss (the name doesn't belong to a module
but to an alias instead).

This function is particularly useful for bp2build, which does not use
real variants when evaluating the blueprint graph.

Test: With soong change, USE_BAZEL_ANALYSIS=1 m libc
Change-Id: I72b8335b642ed2d05e0a38e448cd380acc8d65b0

*build/soong updated to 2f1102fe589714513c5c50db9cfb6a7b4049924e
Merge "Remove bp2build deps mutator"