Autoupdate of 2 submodules
*tools/base updated to 812eee318364fac3bf07de6bc4c0ba7b87092830
Avoid unzipping the repo for AGP

Instead create a bazel java_binary that extracts the repo
for the gradle build to consume.

Bug: 143864616
Test: Manual
Change-Id: Id269e91e25dd04506f04199d1fed0dd128266950

*tools/buildSrc updated to efe4c40a39bef4cd61dff191da3053b6a87106cd
Support config caching with bazel prebuild step

We need to create a task to ensure that the step always runs,
even when the build is from configuration cache.

This change introduces a task and makes all of the java and kotlin
compilation tasks depend on it.

To verify manually:

 1. Run gradle :pAGL with -Phybrid-build=true

 2. Delete a source file built only with bazel that is used in things
    built with gradle, e.g. Nullable.java

 3. Run the gradle build again (verify from configuration cache).
    Without the change the build would incorrectly pass. With this
    change, the build will fail as a required class is missing.

Bug: 143864616
Test: Manually, see above.
Change-Id: Idd07219c232f64d3de7a3e3acc7740f11c0695ec