Autoupdate of 5 submodules
*tools/adt/idea updated to 826a2028e01fc79ddef5ac7378e94eff3cd49e6a
Split maven_library in two rules

This change pulls out the kotlin rule into its own rule,
and makes the maven_library take the library it wraps
as an attribute.

This does not change how maven_library is used, as it
will use the wrapped library provider to behave as a
JavaInfo rule itself.

Test: existing
Bug: 190268212
Change-Id: Ieb56d8a1f179c7ee4d068bdaaab31d6cbe96c17e

*tools/vendor/adt_infra_internal updated to b9bd67ab0bf1927c3777917c155788f9d1eaece1
Split maven_library in two rules

This change pulls out the kotlin rule into its own rule,
and makes the maven_library take the library it wraps
as an attribute.

This does not change how maven_library is used, as it
will use the wrapped library provider to behave as a
JavaInfo rule itself.

Test: existing
Bug: 190268212
Change-Id: I6818199bb002b1871ba55bafaf18ff9ec138e91b

*tools/analytics-library updated to fa0fba2299a8c86e56dc30aac0b953111dca32c0
Split maven_library in two rules

This change pulls out the kotlin rule into its own rule,
and makes the maven_library take the library it wraps
as an attribute.

This does not change how maven_library is used, as it
will use the wrapped library provider to behave as a
JavaInfo rule itself.

Test: existing
Bug: 190268212
Change-Id: Id80ffb4793a66b18250851d79231e6975e350961

*tools/base updated to d447049d857fb3fd1369103bbe63723d78283008
Split maven_library in two rules

This change pulls out the kotlin rule into its own rule,
and makes the maven_library take the library it wraps
as an attribute.

This does not change how maven_library is used, as it
will use the wrapped library provider to behave as a
JavaInfo rule itself.

Test: existing
Bug: 190268212
Change-Id: I224aee03e4efb816091032867a50c02baf3e87f8

*tools/vendor/google updated to d36ff343c80cddacf73252f64085e359f484738d
Split maven_library in two rules

This change pulls out the kotlin rule into its own rule,
and makes the maven_library take the library it wraps
as an attribute.

This does not change how maven_library is used, as it
will use the wrapped library provider to behave as a
JavaInfo rule itself.

Test: existing
Bug: 190268212
Change-Id: Ia3d32dcde04de3c10ae1774cd45702da98252d14