Autoupdate of 2 submodules
*tools/adt/idea updated to 79e1385f948927afd0703554313368b3622a37b5
Remove "Preview" from SdkVersionInfo Api 31

"Preview" was hard coded into version info, and
so will be used even when Api 31 packages are no
longer in preview. Instead, "Preview" is appended
to Api string if the package has a codename
provided.

Test: SdkVersionInfoTest.java
Bug: 170625447
Change-Id: I0ada4053cb24e0d50c0cc069a0ca138de6aecf02

*tools/base updated to 25ce6d373e1ca40fe761584b63a8519b108cded1
Remove "Preview" from SdkVersionInfo Api 31

"Preview" was hard coded into version info, and
so will be used even when Api 31 packages are no
longer in preview. Instead, "Preview" is appended
to Api string if the package has a codename
provided.

Test: SdkVersionInfoTest.java
Bug: 170625447
Change-Id: I1caf5688b359e0ef8c14585cdd74bf6298510cca