Autoupdate of 2 submodules
*tools/adt/idea updated to cb4d8ae3a911f1dec89aef05feb2d0df6812f893
Register lint check for @Discouraged

This change registers the lint check that flags on the usage of APIs
annotated with @Discouraged.

Bug: 178477913
Test: - Presubmit
      - Change was auto-generated by LintInspectionRegistrationTest.kt
Change-Id: I34c9ffe5cdf8f4db4b5d19c2050746e09aac45e1

*tools/base updated to 1644af2fb7a7b507e89b5e2023d87532b2dca974
Add a lint check for the @Discouraged annotation

This change adds DiscouragedDetector, which looks for and flags the
usage of APIs marked @Discouraged. A message with an explanation of why
an API is inadvisable should be defined in the annotation's metadata.

Bug: 178477913
Test: Ran the unit test included in the change.
Change-Id: I133f2d56a3b0329738c8b867443bc6bb63f61275