Autoupdate of 2 submodules
*tools/adt/idea updated to aac74c232cb177b767423f486c464686f321c6b7
[AGP DSL] Decorate variant dimensions

This now uses an insertion-order preserving collection hence the change
in the model snapshot.

Bug: 193452960
Test: Updated Ide[V2]ModelSnapshotComparisonTest expectation files
Change-Id: I6b256d1ed1cab92fc1e1aa9355cf6c40e2c896c6

*tools/base updated to 894fee68ff1d7043bec0299a797f6f889046849d
[AGP DSL] Decorate variant dimensions

This change decorates
 * DefaultConfig
 * BuildType
 * ProductFlavor
 * SigningConfig
with the AGP dsl decorator.

This allows locking, managed properties etc, and includes locking the
NamedDomainObjectContainers. Flavor dimension is also a special case, as
that is modified by AGP after the DSL is locked, that is encoded so the
main property can still be locked.

Does not include migrating blocks within the variant dimensions to also
be decorated, that will follow.

Bug: 193452960
Test: Added new cases to KotlinDslTest
Test: Updated undecorated API expectation file
Change-Id: I9cfa3823cd0e5eb28c2e78da3501589e1b71a3df