Autoupdate of 2 submodules
*tools/adt/idea updated to 61d7759ed4959a2a0fc2641f7ac0d75bf295d955
No hardware accelleration -> fall back

The platform relies on the hardware renderer to provide rendering update
callbacks and SKPs, so fall back to legacy mode if we're not hardware
accelerated.

Test: Added some tests
Bug: 193690457
Change-Id: Ide60ffbdf367f536d3d1ae8bd8b7014089ccc2a6

*tools/base updated to 42fa6774135c428bbad8417ab28536885ac7dff5
No hardware accelleration -> fall back

The platform relies on the hardware renderer to provide rendering update
callbacks and SKPs, so fall back to legacy mode if we're not hardware
accelerated.

Test: Added a test
Bug: 193690457
Change-Id: I8e9c7b833368b8eeae567c1d655a82d389ad24ca