Autoupdate of 2 submodules
*tools/adt/idea updated to 7d6406fb2bfc5f16d5e86024b076479d20583346
adapted snapshot files with new task name.

Task name has changed with the introduction of the anchor task
in Change-Id: Ifdb1513aab383b3a2a7db94071f2088b238cb891

Test: these are tests.
Bug: 69014026
Change-Id: If1d6294691be0a23f2a903f9a191bcefb61a7363

*tools/base updated to 98f96e2a0fa9ea1b3a2e259c7f8f7e3cea87bde4
added an anchor task for the ApksFromBundle task.

Although not technically necessary since APKS_FROM_BUNDLE artifact
type is not public and therefore cannot be replaced by third party
plugins or code, I decided it was more prudent to add an anchor
task in case we make it public later or we decide to replace that
artifact internally depending on flags/etc...

The model stil contains the original listing files, will switch
those in the next CL.

The anchor task is a real task as I wanted to have a type safe
naming in place since the task is in the model (no String mangling).

Test: Existing.
Bug: 69014026
Change-Id: Ifdb1513aab383b3a2a7db94071f2088b238cb891