Autoupdate of 2 submodules
*tools/adt/idea updated to 90cb2a186aacca47f0ecdb2cb8290877623a3052
adapted all snapshot files to model change.

as part of Change-Id: I287a0ddd3231dd77537bfde81e96cd85c4a46b60
the model will not contain the listing file any longer but instead
the redirect file pointing to the listing file. These snapshot
files need to be updated to reflect that change.

Test: Existing.
Bug: 69014026
Change-Id: I3d0355fedb9608dacf44f26cd07db04972ba057e

*tools/base updated to aac3a41059d8f30d98c7d84d529b3b7e6b2e77e7
Changed Model to use the redirect file.

it should be transparent to IDE that the listing file is now
a redirect file.

Test: Existing.
Bug: 69014026
Change-Id: I287a0ddd3231dd77537bfde81e96cd85c4a46b60