Autoupdate of 2 submodules
*tools/base updated to f7bba74cdb62982f0435ec7430a44ebd0c7be4e5
Use Kotlin api and language version 1.4

Update to api and language version 1.4 for Kotlin
as Gradle now ships with Kotlin 1.5. Updating
Kotlin to 1.5 for our build will be done as a
separate change.

Bug: n/a
Test: existing
Change-Id: I97794807a51588265df8b23e7651008684b4dffd

*tools/buildSrc updated to fa1f526da33709900803ff2f15c125fdad906454
Use Kotlin api and language version 1.4

Gradle now ships with Kotlin 1.5 so it is safe to compile to
version 1.4. Update of our build to Kotlin 1.5 will be done
in a separate change.

Bug: n/a
Test: existing
Change-Id: If024d0eeaed046b3bdf2cdf1fc1d78729c18cdc3