196881523: SIOOBE in SourceTransformationTestMode.performEdits
The source transforming test modes visit the AST and then record edits
to be made to the source code at various source offsets.

One challenge here is that some AST concepts are actually duplicated in
the AST from the same source element; this is the case for example for
a property which is inlined in the AST into a separate getter, setter
and field, all mapping to the same source element. If we attempt to
record an edit for each of these, we'd edit the same source region more
than once. For that reason, test modes affected by this generally track
the source PSI elements they have processed.

However, this was not done for the convert-body-to-expression test
mode, which has a similar problem dealing with return expressions
inside companion objects marked @JvmStatic, since here the method body
is duplicated.

This CL fixes this, and more importantly adds a safeguard in the test
machinery making sure that the list of edits coming back from a source
transforming test mode has no overlaps. (The test infrastructure is
already checking for overlaps when attempting to merge multiple
transformations, but it was assuming that an individual test mode has
no overlaps; this is now enforced.)

Test: Unit test included
Fixes: 196881523
Change-Id: I7e2efd36072ab4e862fda9b7dccb3956b19689ce