Autoupdate of 3 submodules
*tools/adt/idea updated to 20a26aeb0030178654b59ee5e6fb2144e52fb457
Kill client if it doesn't progress

We now keep track of how much progress the client has made during
initial connect, and kill it if it doesn't advance for a while. We also
send a metrics event including the stage to which the connection got.

Test: Added/updated tests
Bug: 192685569
Change-Id: Ia3e43d9134f07fcea85f4798014ce3d2d2d95abc

*tools/analytics-library updated to bba83bda50a3e7969a2ce848f88719d4e93a5cd9
Update studio_stats.proto

Logs-Approval: cl/391872209
Bug: 192685569
Test: Added in other changes in topic
Change-Id: Id519be84db01edadfff6170431ec1c3a6c022c58

*tools/base updated to 3c9264e2afa2ab003fec3ee1ca9b234814addb0c
Add progress checkpoint events for metrics

Test: Added/updated tests
Bug: 192685569
Change-Id: I6000d5e46170ed959ad78caa76a0c165c07f8f71