Autoupdate of 2 submodules
*tools/data-binding updated to 236fd6f46dbf617572a67be2340bf8557471a67b
Fix sdk file loading

The api-versions file was moved. It was added to the new location in
platforms in API 26 was removed in platform-tools 31+, in order that it
is versioned alongside with the platform.

Previously, the SDK location would be an argument to the data binding
compiler, with this and the corresponding change in the Android Gradle
Plugin, the the api-versions.xml for the current api level is directly
injected, which means that it can be treated as an input from Gradle's
perspective, and that the build won't be broken again by SDK component
updates.

AGP does not fall back to looking for the api file in platform-tools,
instead the existing bundled version in

Bug: 196847367
Test: SdkVersionTest
Change-Id: I849289c7d41672ca5caf8c4250e04e35b4ce39ae

*tools/base updated to c3f0081c667ead9cee4ed4d59c5bc9f966431b8a
Data binding: Pass api-versions.xml from AGP

Before this change the data binding compiler was finding the
api-versions by hardcoding the path in the SDK in platform-tools.

The api-versions file has been moved to platforms, so it is versioned
now. (There was some overlap, so this is avaliable in platforms since
platform API level 26, and was removed in the platform-tools several
years later in 31.x)

Data binding will fall back to a fixed api versions file shipped as a
java resource in the unusual case of compiling against an older version
than 26.

This fixes the linked issue where data binding currently fails if
platform tools is updated, and removes any implicit dependency on the
SDK state.

Bug: 196847367
Test: Included unit tests
Change-Id: I5782f91a0e73dbcc4f7976ebfa788789bdd933a9