Use doNotShowBuildOutputOnFailure in instrumentation tests
Before this patch, AndroidGradleTaskManager always sets it
false so the build tool-window will always steals a user
focus when a build fails. This is problematic when you run
an Android instrumentation test because Android Studio
now delegates test execution to Gradle and any test failures
are considered as a build failure. See the attached bug
for more details.

Bug: 197247255
Test: GradleConnectedAndroidTestInvokerTest
Change-Id: I25303ff95e10497a1979043bdf48c53655083086