Autoupdate of 2 submodules
*tools/adt/idea updated to 0fc818b0edbd4abd049d26136aa4bf8e76c412db
Update for new model v2 API

Bug: None
Test: existing tests
Change-Id: Idf44170b3b51a9c2ef4dc77ac465d16422657399

*tools/base updated to c44633218c078d4b3f6293ef67e1a0a5e8d12cc8
Fix model support for test-fixtures

getAllArtifacts used ComponentIdentifier as a key to uniquely
identify dependencies but this is not valid when multi-variant
publishing is takend into account (which test fixtures use).

We now need to take into account the capabilities as well, but
not the attributes as they are tied to the query because of
artifactType. The graphs in the nodes are guaranteed to be unique
with just componentIdentifier and capabilities anyway as attributes
are only used for the selection.

This also reflected some issues in the full graph builder where
ResolvedVariantResults again cannot really be used as a match
due to attribute discrepancy, so this code re-uses the same key
as above.

Added 4 integration tests using either android
or java project with test-fixtures, as well as external
java/android libraries with text-fixtures.

Built a small infrastructure to create a dependency graph
and output both a list of ResolvedArtifact and a graph
of DependencyResult. For now this behaves like Gradle
and does not include the capabilities inside the ResolvedArtifact
instances.
This part also make some additions and changes to the existing
Fake implementations of some of the Gradle types used.

Created a few tests based on this infrastructure which
showed that reusing GraphItem instances didn't work, so
this was fixed in FullDependencyGraphBuilder as well.

Finally, removed maven coord cache from ResolvedArtifact as
it was really redundant to have it in so many instances (when
the graphs are large). This reduced quite a lot of injections
of said instance in many other objects

Bug: N/A
Test: New unit and integration tests
Change-Id: Ie2f75bc6b7065c88c62b9838399d72a380675917