Autoupdate of 3 submodules
*tools/data-binding updated to bd4b7af957d783fec358260cb2e370e16b0c35c8
Add the correct stdlib dependency

... for maven_library rules. maven_library was adding
a non-maven dependency for kotlin-stdlib, causing it
not to be declared in the pom.

Some artifacts were adding this dependency manually
which then meant having two dependencies in the
class path.

This change fixes that, and removes the need of
specifying the dependency by hand.

Test: existing
Bug: 190268212
Change-Id: Ibae28f563b5ccb95029cd2bc77a5fcb50bf8fb82

*tools/analytics-library updated to ec975028307fc030d958feceb090b392a7aed5f7
Add the correct stdlib dependency

... for maven_library rules. maven_library was adding
a non-maven dependency for kotlin-stdlib, causing it
not to be declared in the pom.

Some artifacts were adding this dependency manually
which then meant having two dependencies in the
class path.

This change fixes that, and removes the need of
specifying the dependency by hand.

Test: existing
Bug: 190268212
Change-Id: Id0b74834924f7e677475f653e9a312d286ed43c4

*tools/base updated to c2bbbec5b92ca0eed9f44b8655cac5b0aea64e0b
Add the correct stdlib dependency

... for maven_library rules. maven_library was adding
a non-maven dependency for kotlin-stdlib, causing it
not to be declared in the pom.

Some artifacts were adding this dependency manually
which then meant having two dependencies in the
class path.

This change fixes that, and removes the need of
specifying the dependency by hand.

Test: existing
Bug: 190268212
Change-Id: I1203e1af5a06587ecbfbfefa24c28917c985d7ac