Autoupdate of 4 submodules
*tools/data-binding updated to a946fc71f32d52f9395bd291aba8f0852c9e3ad1
Update to Compose 1.0.1

Bug: 194823927
Test: Existing tests pass
Change-Id: Ia484d58a5a6677c281f8198bd457b808d3ac4c2e

*prebuilts/tools updated to 5c38cadc8c3318253c41fbb06f32dee0d88643b7
Update to Compose 1.0.1

Remove Compose 1.0.0 and upgrade Kotlin to 1.5.21, required for Compose.

Bug: 194823927
Test: Existing tests pass
Change-Id: I57293630c5171e4ff269712d3c01e20e0846eabf

*tools/adt/idea updated to a4623f2edc46effea14acff8c8e721005f28411b
Update Compose to 1.0.1

This includes updating Kotlin to the required 1.5.21 for Compose
templates. This CL also removes Compose 1.0.0 and Kotlin 1.5.10
references for the templates.

Bug: 194823927
Test: Existing tests pass
Change-Id: I0cedc8e2c9f7cd58d687782e9c26746afd2b9770

*tools/base updated to 4fcb44f4786fa5925011cb749d7fb053e7cf255c
Update Compose to 1.0.1

This includes updating Kotlin to the required 1.5.21 for Compose
templates. This CL also removes Compose 1.0.0 and Kotlin 1.5.10
references for the templates.

Bug: 194823927
Test: Existing tests pass
Change-Id: If739b3e62af0d9095ed26f29842a0358d52f8122