Autoupdate of 2 submodules
*prebuilts/r8 updated to 5a9880a1bd63acf1aaf3faef2a608abbaa85e53d
Update D8 R8 to 3.1.17-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.1.17-dev
Test: ./gradlew check
Bug:

Change-Id: Idb830710050be6d54bac34307597ffb97c65eafb

*tools/base updated to 87916293277dc86e1f49278c5020513e99ac4a9e
Account for outlining in testBytecodeOfFunctionWithDesugaredLibraryParam

Bug:
Test: ./gradlew check
Change-Id: Ibde944b581e66ad13314bf0ace58f0ba9b55148c