Autoupdate of 3 submodules
*prebuilts/tools updated to cf0c82b113d8f74014993daa12063c21bfa4dbee
Add artifacts for adt_infra_internal

Bug: 190268212
Test: //tools/vendor/adt_infra_internal/...
Change-Id: I1272e4a6ce1c2d365716f7d5ec59ce793a449fa8

*tools/vendor/adt_infra_internal updated to 95852c7df5fd34fe81c2a8a8db5071eb78a62368
Use new maven rules for adt_infra_internal

This affects the versions of the following dependencies
used by the RBE client:

  1. protobuf: old=3.9.0, new=3.10.0
  2. grpc:     old=1.24.0, new=1.21.1

This version change is because there can only be one
"resolved" version for each of these dependencies in @maven.

I expect protobuf upgrade to not cause any problems. But, I'm
not sure about the grpc downgrade. If the RBE client does not
work with a downgraded grpc, we can consider
  1. creating a java_import rule for grpc 1.24.0,
  2. upgrading the rest of the Studio to use grpc 1.24.0.

Bug: 190268212
Test: //tools/vendor/adt_infra_internal/...
Change-Id: I61d7819fe2017b7b84b22028867d609055c46ce4

*tools/base updated to 053ebcc4052292acc582309b1b6d774e26c1c1da
Add artifacts of adt_infra_internal to @maven

Bug: 190268212
Test: //tools/vendor/adt_infra_internal/...
Change-Id: I1303493669e6d8ec3683d1b934dea3be63d316bc