Autoupdate of 2 submodules
*prebuilts/tools updated to a0ac5b99ad32be6760767a5210200865ed55e8eb
Convert usages of maven_java_import to java_import

The "maven" capabilities of this rule was already removed,
and the "license" capabilities were not being consumed by
anyone. This leaves it with just the Java capabilities.

Bug: 190268212
Test: presubmit
Change-Id: I8ebea57efa902ff3f37723b46f836c0bbe5f9b72

*tools/base updated to 213bb9dfc9fc4dc8a1e47b17d356dd2b474717cb
Remove maven_java_import rule

The few remaining usages can be converted to java_import.

Bug: 190268212
Test: presubmit
Change-Id: I786e37ba61fd5d258237ef10208d99bad8041a31