197966892: LintFix's annotate function misplaces annotations
Quickfixes typically do not require a specific range to be set on them;
instead, they'll use the location range for the associated incident.
However, they *do* have a Location property which you can set if a fix
applies to a different location than the warning.

Lint will also pick a "suitable" location when you do not specify an
explicit location type. For example, for a method, it will pick the
method name identifier instead of the entire method body.

However, this means that if you have a quickfix and *don't* set an
associated range, the quickfix will think the range is just the
identifier. Some quickfixes will not work correctly in this case, such
as the annotation quickfix.

This CL improves this situation by recording the cases where the
location range is narrowed in this way, and automatically updating any
associated quickfixes with the initial location range instead. It also
fixes a few other misc location related issues.

Fixes: 197966892
Test: Existing and new unit tests
Change-Id: I31f5b35c1d9bd07622c359a5b48a228b9e4689ca