Autoupdate of 2 submodules
*tools/adt/idea updated to 96a035f2eb66993a68dc8c9c9eedeceac3a4ced9
Map composite build IDs in models vs

Bug: n/a
Test: n/a
Change-Id: I68267d5104386617b1a192b99e0e877827675dab

*tools/base updated to e7c000517eb155ec09b53fcb0c14e5f621d0cf35
Fix v2 model for better composite build support.

In order to properly have a unique identifier on the
composite builds, this change also adds a build ID
map to the model. The build ID map is (name, rootdir)
which allows us to reconcille the API in the model
action which only has rootDir with the information
coming from the dependency model which only has name.

This also include new model tests for
composite builds, using a new  infrastructure
to create such-projects

Bug: N/A
Test: new tests
Change-Id: If79b7626b02e08db3013f4b5228d1d540b361da0