Update D8 R8 to 3.1.19-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.1.19-dev
Test: ./gradlew check
Bug:

Change-Id: Iea0362f91cf35f3d560325043d3212406c734f82