Autoupdate of 4 submodules
*prebuilts/tools updated to 7d6fecde858d1d31fd9f48f8de390bb269b4200c
Remove protobuf versions 3.0.0 and 3.4.0

Bug: 198494349
Test: presubmit
Change-Id: I117c7bb89b204693c49edb943dc0d18452f3c236

*tools/adt/idea updated to f6400f8af898bf25f1b4281699c1cc1f1ec68ad1
Update protobuf in studio to use default version

...which is currently 3.10.0.

The goal is to reduce the number of protobuf versions to,
preferrably, only one version.

Bug: 198494349
Test: presubmit
Change-Id: Ica12ebc65ead6e898844b8e3f6ffe5828dcccbb7

*tools/analytics-library updated to e02a2754f931fe960829db7fbfe1e02cb1146095
Update proto in analytics-library to use default

...which is currently 3.10.0

The goal is to reduce the number of protobuf versions to,
preferrably, only one version.

Bug: 198494349
Test: presubmit
Change-Id: I665ec94ed3a95de731767a3fdabcdcb552ad3268

*tools/base updated to 882c06f1dad17675e60a88f20c737363e425b4cc
Update protobuf in ddmlib to the default version

... which is 3.10.0

Bug: 198494349
Test: presubmit
Change-Id: Id2b298619c465a7143bd198958dc5f1481c20976