Improve msg for compile target incompatibility
when using AAR Metadata.

This also introduces BaseTask.projectPath as project.name
is only the last segment of the path and therefore not using.
The build should always use the path rather than the name,
and this change deprecate the name in favor of the path.
A later commit will remove usage of projectName.

A few tasks where already defining this so this commit
changes them to use the new base property.

Test: existing tests
Bug: 197553058
Change-Id: I4ac1037501e5d79d207f1047d1fe0f2c7214673b