Autoupdate of 3 submodules
*build/make updated to f999a9d145c89cd93dd2004830aba5b5a53b0aaf
Merge "Remove soong_docs from droidcore."
*build/blueprint updated to 5d5ed5fc6b6f6c192d2337d5592bcd31773cee02
Remove PrimaryBuilderExtraFlags and cull Args.

PrimaryBuilderExtra flags was only used from soong_ui and has nothing to
do with Blueprint files or modules and Args contained a lot of members
that were likewise only used for plumbing withing soong_ui or for
command line processing for soong_build.

Test: Presubmits.
Change-Id: Ibf009e446b4044309a331592a88a4aa86b4801e3

*build/soong updated to 89fcdcb78813836bb73d5bee45018ad6cfe78677
Refactor the creation of soong_build calls.

They were a bit repetitive.

In addition, make the environment checks more correct; they could
probably use an integration test or two to make sure that when the
environment changes, exactly those outputs are rebuilt that need to be,
but for now, this is an improvement already.

Test: Presubmits.
Change-Id: Idd79b81ca6975d57d00e5bf4699d266152505ff8