Autoupdate of 2 submodules
*tools/adt/idea updated to 6abef25f31379acfeb4392eed6c48afeaa852436
Attach long field values as files instead

Crash has a limit on the number of bytes for fields in the request body. This change sanitizes fields to ensure Crash doesn't exclude information from reports. Fields that are too long for Crash are truncated and the original raw text is included as a file.

For the future, we intend to write a wrapper around MultipartEntityBuilder to abstract this logic and pass around this object in the serialization methods for reports.


Change-Id: I9ec80bc56d6255f96e3bb6776254508125d10aeb
Bug: 196853911

*tools/analytics-library updated to fceafc866445d0f5ef393a78af4c9908ccca6273
Attach long field values as files instead

Add utility function to sanitize request body fields that are too long for Crash

Change-Id: I0ea75d08b5dfa7432ebe2e7571a918e23a53d257