Autoupdate of 2 submodules
*tools/adt/idea updated to a7a02bbbb6c27b6cecae728a4e1159a21bec4907
revert some golden files changes

the APKs will only be located in intermediates only if the APKs are
optimized for specific ABI, density or device API.

If the build is invoked from IDE but is not optimized, it will stay
in outputs/ like before.

Test: this is a test
Bug: 198501473
Change-Id: I0ddc10f9465c20b015e83924d25b4dc706c981d8

*tools/base updated to 51f22d4afe9d53cb4f9d19b1974e1a208ba08263
only move apks to intermediates for custom builds

APK files were produced in intermediates when invoked from the
IDE. However, when doing a "Build projects", non optimized builds
are requested and should therefore kept under outputs folder.

Ignore the INVOKED_FROM_IDE flag and only use the target abi,
target density and device API as signal for optimized builds
that should be placed in intermediates.

Bug: 198501473
Test: added unit tests
Change-Id: I7bd9f20183cf2df5f07cf8bcba55700c6330e12b