Autoupdate of 2 submodules
*tools/adt/idea updated to dd7dce3277adc34bbf7ea2a263cb781317c49ba8
Update model v2 conversion

since GlobalLibraryMap is gone and the data
is now part of VariantDependencies.

Bug: N/A
Test: existing tests
Change-Id: I42c038c2129e9351e4135da0b1b99551bad48acb

*tools/base updated to 4852757510235fa30defe4cc574840e1ec0f9fee
Merge GlobalLibraryMap into VariantDependencies

With upcoming caching on a per-project basis for
the sync model, we cannot have a global model that's
queried on a random project.
Because of this, this change moves the library map
to be computed and returned on a per-project basis.

It's still advantageous to keep the GraphItem as
light-weight as possible, and remove redundancy
across the graphs for the main, test, androidTest and
fixture artifacts.

This change also renames GlobalLibraryBuildService to
GlobalSyncService since it's not tied to the Library
instances anymore but still provides services to sync.

Finally, tweak and improve the test fixtures for
querying model and comparing them.

Bug: N/A
Test: existing tests. Updated the snapshots.
Change-Id: I2c5d9a53670e90e92e5122fbc343ccb598fb488f