Allow disabling JNI refernce search
For some projects, JNI refernce search can be very slow. Though the CIDR
logic already checks cancellation, somehow it can still cause UI
freezes. Hence, as a temporary fix this change add a registry to disable
JNI refernce search.

A registry is used instead of a studio flag because the former is easier
for users to control, while the latter needs user to starts Android
Studio with a commandline flag.

Bug: 197323118
Test: manual
Change-Id: I7957e1b4d82b50fbc8a0878e397cf256d9c37e3e