ART: Enables Debugging with LLDB for dex2oat.
Currently, there is not a flag present to enable debugging the dex2oat
compiler on host. Therefore, because we want to be able to debug the
dex2oat compiler when testing, this patch adds two flags,
"--gdb-dex2oat" and "--gdb-dex2oat-args". As a result, when the
"--gdb-dex2oat" flag is present in a test command, it will
run LLDB on dex2oat.

Note that to be compliant with the current naming convention,
the relevant variable names include the "GDB" prefix instead of "LLDB".

Additionally, we can pass arguments that are delimited by a semi-column
to LLDB with "--gdb-dex2oat-args", e.g.
"--gdb-dex2oat-args='-X'".

Note that this patch does not enable dex2oat debugging on target.

Test: art/test.py -v -j12 --host --64  -t 554-checker-rtp-checkcast\
      --run-test --optimizing --gdb-dex2oat

Change-Id: I392a57a41a90550cc5dc766ee9f148327a9c21d2