[Dsl] Move some GradleDslFile fields
The concept of parent and child files only applies to build files, not
to generic GradleDslFiles; it is basically a mirror (maybe one of those
deforming ones) onto the hierarchical Gradle project model.  In any
case, it is almost always wrong to query a settings file (or a
properties file) for its parent or children, so move the parent and
child fields into GradleBuildFile from the parent GradleDslFile class.

This rearrangement causes a few knock-on effects throughout the Dsl
model implementation, most of which are straightforward moves or
renames.  At one or two points, primarily in resolution or the related
operation of computing an external name for a property, it becomes
necessary to test explicitly for GradleBuildFile.

The propertiesFile field is likewise only appropriate to check as we
currently do from GradleBuildFiles; there is no concept of a "sibling"
project properties file from the settings file, for instance.  Move the
propertiesFile field to GradleBuildFile too.

Bug: n/a
Test: existing (no change in functionality)
Change-Id: I7dbeecc605bc1d769a21a1bb0d0541854fb7e9e4