Autoupdate of 2 submodules
*tools/adt/idea updated to da9d2ea4de72cffa34165ab0bf84556d2ca5d6eb
[AUA] Remove buildToolsVersion if it would be ignored

New AGP (by which we mean "since 3.0" :-) will ignore a too-low
specification of buildToolsVersion, replacing it with its minimum
supported version with a warning.  We don't have a convenient table of
data to support removing buildToolsVersion in all cases, but for
upgrades to AGP 7.0 and 7.1 we can remove buildToolsVersion if it will
be below the minimum version supported by those AGPs.

Fixes: 183650762
Test: added
Change-Id: Idb34b09174e7344262cd956d75ff909607f00073

*tools/analytics-library updated to 42e86ce07791c1a8a743863ad9b053b8a7b58663
Update studio_stats.proto

Logs-Approval: cl/398000132
Bug: 183650762
Test: existing
Change-Id: Idccdff70401890f113333fe05773829408238681