Autoupdate of 3 submodules
*frameworks/base updated to e731e209d326b8e531d85a0deae611689e0afa3b
Merge "ServiceManagerNative: add getConnectionInfo"
*frameworks/native updated to c2b54f6f547fc292d2e896f361bc1188cf02faeb
Merge "Add getConnectionInfo API to service manager"
*system/libvintf updated to 1b5e0e7ef71efac8b6db8b4a260c3f60d16a0743
Add support for new ip and port fields for AIDL hals

This is only supported for AIDL HALs.
Allows the device to find remote HALs listening on that static IP
address and port number.
This is done by new attributes in the transport element in the manifest
files for hals:
   <transport ip="1.2.3.4" port="1234">inet</transport>

Test: atest libvintf_test
Bug: 198207801

Change-Id: I9f013e5e88c959e2b96e2631fa50107cf2ad2db8