Autoupdate of 2 submodules
*tools/analytics-library updated to c20054517ef27864ab6b7923fc387f8c61e0c24b
[Hybrid build] Add pom names and descriptions

That were lost in the switch to bazel.

The ones we actually ship are covered by GmavenZipTest,
so all will be verified once all of analytics-library
is built with bazel.

Bug: 143864616
Test: Updated GmavenZipTest
Change-Id: Iedea24bd765dcbeafc01ff3f95d59b229642dd1b

*tools/base updated to e513281f3da49abc2f82e4f0304ca65aabe64e2a
Add pom names and descriptions for ddmlib and analytics

That were lost in the switch to bazel.

The ones we actually ship are covered by GmavenZipTest,
so all will be verified once all of analytics-library
is built with bazel.

Bug: 143864616
Test: Updated GmavenZipTest
Change-Id: Ic5a4070ba25006c7195d49c307f66a0ec45d5c3c